network layer changes: implemented tcp-style retransmit timeout. made bandwith limit dynamically adjustable, roughly based on tcp congestion avoidance/handling algorithms, minus slow-start. this might need some tuning, as most ss traffic is unreliable, so congestion is harder to detect. also made bandwith limit use smoothed average.
added a millisecond version of GTC to util to get more accurate timing
information for the new network layer stuff.

removed SetLimit from net.

modified the file transfer cancel request handler. it doesn't work
perfectly with cont yet.

fixed missing interface gets in admincmd.

fixed some declarations of time values.

made keepalive ping packet reliable.

bumped version number, updated news.txt.

+++ src/game_timer.c
@@ -187,7 +187,7 @@ local void Ctimereset(const char *params, Player *p, const Target *target)
 {
 	Arena *arena = p->arena;
 	timerdata *td = P_ARENA_DATA(arena, tdkey);
-	long gamelen = td->gamelen;
+	unsigned gamelen = td->gamelen;
 
 	if (gamelen)
 	{
